+++ src/redux/sagas/user.saga.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { put, takeLatest } from 'redux-saga/effects';
 
-const handleErrors = (msg, err) => {
+export const handleErrors = (msg, err) => {
   console.log(msg, err);
   alert('Error fetching data:' + msg)
 }
@@ -65,74 +65,14 @@ function* fetchPlayedList() {
   }
 }
 
-function* addToWishlist({ payload }) {
-  try {
-    yield axios.post('/api/wishlist', payload);
-    yield put({ type: 'USER/FETCH_WISHLIST' })
-  } catch (err) {
-    handleErrors('Adding game to wishlist failed', err)
-  }
-}
 
-function* deleteFromWishlist({ payload }) {
-  try {
-    yield axios.delete('/api/wishlist/' + payload);
-    yield put({ type: 'USER/FETCH_WISHLIST' })
-  } catch (err) {
-    handleErrors('Deleting from wishlist failed', err)
-  }
-}
 
-function* addToIgnorelist({ payload }) {
-  try {
-    yield axios.post('/api/ignorelist', payload)
-    yield put({ type: 'USER/FETCH_IGNORELIST' })
-  } catch (err) {
-    handleErrors('Adding to ignore list failed', err)
-  }
-}
-
-function* deleteFromIgnorelist({ payload }) {
-  try {
-    yield axios.delete('/api/ignorelist/' + payload)
-    yield put({ type: 'USER/FETCH_IGNORELIST' })
-  } catch (err) {
-    handleErrors('Deleting from ignore list failed', err)
-  }
-}
-
-function* addToPlayedList({ payload }) {
-  try {
-    yield axios.post('/api/played', payload)
-    yield put({ type: 'USER/FETCH_PLAYED_LIST' })
-  } catch (err) {
-    handleErrors('Adding to played list failed', err)
-  }
-}
-
-function* deleteFromPlayedList({ payload }) {
-  try {
-    yield axios.delete('/api/played/' + payload)
-    yield put({ type: 'USER/FETCH_PLAYED_LIST' })
-  } catch (err) {
-    handleErrors('Deleting from played list failed', err)
-  }
-}
 
 function* userSaga() {
   yield takeLatest('FETCH_USER', fetchUser);
   yield takeLatest('USER/FETCH_WISHLIST', fetchWishlist);
   yield takeLatest('USER/FETCH_IGNORELIST', fetchIgnorelist);
   yield takeLatest('USER/FETCH_PLAYED_LIST', fetchPlayedList);
-
-  yield takeLatest('USER/WISHLIST/ADD', addToWishlist)
-  yield takeLatest('USER/WISHLIST/DELETE', deleteFromWishlist)
-
-  yield takeLatest('USER/IGNORELIST/ADD', addToIgnorelist)
-  yield takeLatest('USER/IGNORELIST/DELETE', deleteFromIgnorelist)
-
-  yield takeLatest('USER/PLAYEDLIST/ADD', addToPlayedList)
-  yield takeLatest('USER/PLAYEDLIST/DELETE', deleteFromPlayedList)
 }
 
 export default userSaga;
